+++ src/main/java/org/morphix/reflection/Reflection.java
@@ -156,29 +156,13 @@ public interface Reflection {
 	}
 
 	/**
-	 * Returns a class based on a class name.
-	 *
-	 * @param <T> returned type
-	 *
-	 * @param className the {@linkplain ClassLoader##binary-name binary name}
-	 * @return a class based on a class name
-	 */
-	static <T> Class<T> getClass(final String className) {
-		try {
-			return JavaObjects.cast(Class.forName(className));
-		} catch (ClassNotFoundException e) {
-			return null;
-		}
-	}
-
-	/**
 	 * Returns true if the class given by its name is present in the class path, false otherwise.
 	 *
 	 * @param className the {@linkplain ClassLoader##binary-name binary name}
 	 * @return true if the class is present in the class path, false otherwise
 	 */
 	static boolean isClassPresent(final String className) {
-		return null != getClass(className);
+		return null != Classes.Safe.getOne(className);
 	}
 
 }